changed to long long

+++ fast.c
@@ -14,14 +14,14 @@
 
 typedef struct rational
 {
-	long p;
-	long q;
+	long long p;
+	long long q;
 } rat;
 
 rat reduce(rat x)
 {
-	long a = x.p;
-	long b = x.q;
+	long long a = x.p;
+	long long b = x.q;
 
 	if (a == 0 || b == 0)
 	{
@@ -318,12 +318,12 @@ bool fm(size_t rows, size_t cols, signed char a[rows][cols], signed char c[rows]
 		for (j = 0; j < var; j++)
 		{
 			rat ttemp;
-			ttemp.p =(long)a[i][j];
+			ttemp.p =(long long)a[i][j];
 			ttemp.q = 1;
 			t[i][j] = ttemp;
 		}
 
-		qtemp.p = (long)c[i];
+		qtemp.p = (long long)c[i];
 		qtemp.q = 1;
 		q[i] = qtemp;
 	}
